backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langgraph==0.2.20
openai==1.40.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
faiss-cpu==1.7.4
numpy==1.26.0
supabase==2.0.0
python-dotenv==1.0.0
httpx==0.24.1
aiofiles==23.2.1
textblob==0.17.1
spacy==3.7.2
tiktoken==0.7.0
pymupdf==1.23.8
pdfplumber==0.10.3